create different components
Separate into 3 different component and use the main page to call those component.
